+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - bug fixes
     - nested property inheritance (eg. `@{id_@{id_x}}`) not updating properly
-
+    - canvas based widgets not updating when hidden
+    
 ## 1.9.12
 
 - bug fixes